requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.3
Django==3.1.1
django-filebrowser-no-grappelli==3.8.0
django-taggit==1.3.0
django-tinymce4-lite==1.8.0
isort==5.5.2
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
validate-email==1.3
wrapt==1.12.1
